Fix bug in mutable and shared annotations (#292)
When a reconciliation loop adds a new annotation to the map it is
silently reused for other apps reconciling at a later time.

This leads to a case where an app who doesn't have a Prometheus
block specified gets attempted scraped at the previous reconciled
app's (with Prometheus config) port and path.

SKIP-1134