refactor alpha1 and beta1 to use common types
Moved prometheus and status into common types
Created alias files to define the types in both v1alpha1 and v1beta1
packages. Replaced the alias folder with a single alias file so that it
is defined in the v1alpha1 package or v1beta1 package.
Updated the converter from v1alpha1 to v1beta1