🐛 Fix manager setup
After testing the skiperator with argoCD locally we found out that the
skiperator did not actually start the reconcilers. This was due to an
programming mistake, where the manager was started before all the other
reconcilers were initialized. This is fixed now and the skiperator sets
up the reconcilers correctly. For local testing we added argocd to the
exclusion config, THIS MUST BE REMOVED UPON CLEANUP.